pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "APIs for a portfolio"
authors = [
    {name = "Jacques Niyonkuru",email = "niyonkurujacques@gmail.com"}
]

readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "annotated-doc (==0.0.3)",
    "annotated-types (==0.7.0)",
    "anyio (==4.11.0)",
    "certifi (==2025.11.12)",
    "click (==8.3.0)",
    "colorama (==0.4.6)",
    "dnspython (==2.8.0)",
    "email-validator (==2.3.0)",
    "exceptiongroup (==1.3.1)",
    "fastapi-cli (==0.0.16)",
    "fastapi-cloud-cli (==0.3.1)",
    "greenlet (==3.3.0)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httptools (==0.7.1)",
    "httpx (==0.28.1)",
    "idna (==3.11)",
    "jinja2 (==3.1.6)",
    "markdown-it-py (==4.0.0)",
    "markupsafe (==3.0.3)",
    "mdurl (==0.1.2)",
    "pydantic (==2.12.3)",
    "pydantic-core (==2.41.4)",
    "pygments (==2.19.2)",
    "python-dotenv (==1.2.1)",
    "python-multipart (==0.0.20)",
    "pyyaml (==6.0.3)",
    "rich (==14.2.0)",
    "rich-toolkit (==0.15.1)",
    "rignore (==0.7.6)",
    "sentry-sdk (==2.44.0)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "starlette (==0.48.0)",
    "tomli (==2.3.0)",
    "typer (==0.20.0)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.15.0)",
    "urllib3 (==2.5.0)",
    "uvicorn (==0.38.0)",
    "uvloop (==0.22.1)",
    "watchfiles (==1.1.1)",
    "websockets (==15.0.1)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "fastapi (>=0.133.0,<0.134.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.scripts]
dev="app.cli:dev"
seed="app.cli:seed"
